+++ lib/appium_lib_core/driver.rb
@@ -40,7 +40,7 @@ module Appium
 
       def initialize(appium_lib_opts)
         @custom_url = appium_lib_opts.fetch :server_url, nil
-        @default_wait = appium_lib_opts.fetch :wait, Driver::DEFAULT_IMPLICIT_WAIT
+        @default_wait = appium_lib_opts.fetch :wait, nil
         @enable_idempotency_header = appium_lib_opts.fetch :enable_idempotency_header, true
 
         # bump current session id into a particular file
@@ -51,7 +51,7 @@ module Appium
 
         @port = appium_lib_opts.fetch :port, Driver::DEFAULT_APPIUM_PORT
 
-        # timeout and interval used in ::Appium::Comm.wait/wait_true
+        # timeout and interval used in ::Appium::Commn.wait/wait_true
         @wait_timeout  = appium_lib_opts.fetch :wait_timeout, ::Appium::Core::Wait::DEFAULT_TIMEOUT
         @wait_interval = appium_lib_opts.fetch :wait_interval, ::Appium::Core::Wait::DEFAULT_INTERVAL
 
@@ -135,11 +135,9 @@ module Appium
       attr_reader :export_session_path
 
       # Default wait time for elements to appear in Appium server side.
-      # Defaults to {::Appium::Core::Driver::DEFAULT_IMPLICIT_WAIT}.<br>
       # Provide <code>{ appium_lib: { wait: 30 } }</code> to {::Appium::Core.for}
       # @return [Integer]
       attr_reader :default_wait
-      DEFAULT_IMPLICIT_WAIT = 0
 
       # Appium's server port. 4723 is by default. Defaults to {::Appium::Core::Driver::DEFAULT_APPIUM_PORT}.<br>
       # Provide <code>{ appium_lib: { port: 8080 } }</code> to {::Appium::Core.for}.
@@ -187,8 +185,6 @@ module Appium
       # @option opts [Hash] :caps Appium capabilities.
       # @option opts [Hash] :capabilities The same as :caps.
       #                                   This param is for compatibility with Selenium WebDriver format
-      # @option opts [Hash] :desired_capabilities The same as :caps.
-      #                                           This param is for compatibility with Selenium WebDriver format
       # @option opts [Appium::Core::Options] :appium_lib Capabilities affect only ruby client
       # @option opts [String] :url The same as :custom_url in :appium_lib.
       #                            This param is for compatibility with Selenium WebDriver format
@@ -199,10 +195,8 @@ module Appium
       #
       #     # format 1
       #     @core = Appium::Core.for caps: {...}, appium_lib: {...}
-      #     # format 2. 'capabilities:' or 'desired_capabilities:' is also available instead of 'caps:'.
+      #     # format 2. 'capabilities:' is also available instead of 'caps:'.
       #     @core = Appium::Core.for url: "http://127.0.0.1:8080/wd/hub", capabilities: {...}, appium_lib: {...}
-      #     # format 3. 'appium_lib: {...}' can be blank
-      #     @core = Appium::Core.for url: "http://127.0.0.1:8080/wd/hub", desired_capabilities: {...}
       #
       #
       #     require 'rubygems'
@@ -230,7 +224,7 @@ module Appium
       #     @core.start_driver # Connect to 'http://127.0.0.1:8080/wd/hub' because of 'port: 8080'
       #
       #     # Start iOS driver with .zip file over HTTP
-      #     #  'desired_capabilities:' or 'capabilities:' is also available instead of 'caps:'. Either is fine.
+      #     # 'capabilities:' is also available instead of 'caps:'. Either is fine.
       #     opts = {
       #              capabilities: {
       #                platformName: :ios,
@@ -254,7 +248,7 @@ module Appium
       #     # Start iOS driver as another format. 'url' is available like below
       #     opts = {
       #              url: "http://custom-host:8080/wd/hub.com",
-      #              desired_capabilities: {
+      #              capabilities: {
       #                platformName: :ios,
       #                platformVersion: '11.0',
       #                deviceName: 'iPhone Simulator',
@@ -367,11 +361,12 @@ module Appium
         end
 
         begin
-          # included https://github.com/SeleniumHQ/selenium/blob/43f8b3f66e7e01124eff6a5805269ee441f65707/rb/lib/selenium/webdriver/remote/driver.rb#L29
-          @driver = ::Appium::Core::Base::Driver.new(http_client: @http_client,
-                                                     desired_capabilities: @caps,
+          @driver = ::Appium::Core::Base::Driver.new(listener: @listener,
+                                                     http_client: @http_client,
+                                                     capabilities: @caps, # ::Selenium::WebDriver::Remote::Capabilities
                                                      url: @custom_url,
-                                                     listener: @listener)
+                                                     wait_timeout: @wait_timeout,
+                                                     wait_interval: @wait_interval)
 
           if @direct_connect
             d_c = DirectConnections.new(@driver.capabilities)
@@ -412,6 +407,8 @@ module Appium
 
       # Ignore setting default wait if the target driver has no implementation
       def set_implicit_wait_by_default(wait)
+        return if @default_wait.nil?
+
         @driver.manage.timeouts.implicit_wait = wait
       rescue ::Selenium::WebDriver::Error::UnknownError => e
         unless e.message.include?('The operation requested is not yet implemented')
@@ -479,21 +476,6 @@ module Appium
       def platform_version
         p_version = @driver.capabilities['platformVersion'] || @driver.session_capabilities['platformVersion']
         p_version.split('.').map(&:to_i)
-      end
-
-      # Takes a png screenshot and saves to the target path.
-      #
-      # @param png_save_path [String] the full path to save the png
-      # @return [File]
-      #
-      # @example
-      #
-      #   @core.screenshot '/tmp/hi.png' #=> nil
-      #   # same as '@driver.save_screenshot png_save_path'
-      #
-      def screenshot(png_save_path)
-        ::Appium::Logger.warn '[DEPRECATION] screenshot will be removed. Please use driver.save_screenshot instead.'
-        @driver.save_screenshot png_save_path
       end
 
       private
@@ -568,10 +550,7 @@ module Appium
       def validate_keys(opts)
         flatten_ops = flatten_hash_keys(opts)
 
-        # FIXME: Remove 'desired_capabilities' in the next major Selenium update
-        unless opts.member?(:caps) || opts.member?(:capabilities) || opts.member?(:desired_capabilities)
-          raise Error::NoCapabilityError
-        end
+        raise Error::NoCapabilityError unless opts.member?(:caps) || opts.member?(:capabilities)
 
         if !opts.member?(:appium_lib) && flatten_ops.member?(:appium_lib)
           raise Error::CapabilityStructureError, 'Please check the value of appium_lib in the capability'
@@ -592,8 +571,7 @@ module Appium
 
       # @private
       def get_caps(opts)
-        # FIXME: Remove 'desired_capabilities' in the next major Selenium update
-        Core::Base::Capabilities.create_capabilities(opts[:caps] || opts[:capabilities] || opts[:desired_capabilities] || {})
+        Core::Base::Capabilities.create_capabilities(opts[:caps] || opts[:capabilities] || {})
       end
 
       # @private