+++ test/functional/android/android/image_comparison_test.rb
@@ -100,7 +100,7 @@ class AppiumLibCoreTest
 
         assert image_element.inspect
         assert image_element.hash
-        assert image_element.ref =~ /\Aappium-image-element-[a-z0-9\-]+/
+        assert image_element.id =~ /\Aappium-image-element-[a-z0-9\-]+/
 
         el_location = el.location
         image_location = image_element.location
@@ -143,7 +143,7 @@ class AppiumLibCoreTest
 
         assert image_element.inspect
         assert image_element.hash
-        assert image_element.ref =~ /\Aappium-image-element-[a-z0-9\-]+/
+        assert image_element.id =~ /\Aappium-image-element-[a-z0-9\-]+/
 
         el_location = el.location
         image_location = image_element.location
@@ -185,7 +185,7 @@ class AppiumLibCoreTest
         image_element = @driver.find_element_by_image AppiumLibCoreTest.path_of('test/functional/data/test_android_nfc_270.png')
         assert image_element.inspect
         assert image_element.hash
-        assert image_element.ref =~ /\Aappium-image-element-[a-z0-9\-]+/
+        assert image_element.id =~ /\Aappium-image-element-[a-z0-9\-]+/
 
         el_location = el.location
         image_location = image_element.location