+++ test/functional/ios/ios/image_comparison_test.rb
@@ -100,7 +100,7 @@ class AppiumLibCoreTest
 
         assert image_element.inspect
         assert image_element.hash
-        assert image_element.ref =~ /\Aappium-image-element-[a-z0-9\-]+/
+        assert element.id =~ /\Aappium-image-element-[a-z0-9\-]+/
         assert !image_element.visual.nil?
 
         el_location = el.location
@@ -146,7 +146,7 @@ class AppiumLibCoreTest
 
         assert image_element.inspect
         assert image_element.hash
-        assert image_element.ref =~ /\Aappium-image-element-[a-z0-9\-]+/
+        assert element.id =~ /\Aappium-image-element-[a-z0-9\-]+/
 
         el_location = el.location
         image_location = image_element.location