+++ test/functional/ios/webdriver/device_test.rb
@@ -60,7 +60,7 @@ class AppiumLibCoreTest
 
         e = @@core.wait { @@driver.find_element :predicate, 'value == "Placeholder text"' }
         e.click
-        @@driver.set_immediate_value e, 'hello'
+        e.immediate_value 'hello'
 
         # Using predicate case
         e = @@core.wait { @@driver.find_element :predicate, by_predicate('hello') }
@@ -145,7 +145,7 @@ class AppiumLibCoreTest
 
       def test_implicit_wait
         # checking no method error
-        assert(@@driver.manage.timeouts.implicit_wait = @@core.default_wait)
+        assert(@@driver.manage.timeouts.implicit_wait = 0)
       end
 
       def test_rotate
@@ -158,8 +158,8 @@ class AppiumLibCoreTest
 
       # TODO: add an async execuite test case
       # def test_async_execuite
-      #   @@driver.execute_async_script('mobile: tap', x: 0, y: 0, element: element.ref)
-      #   @@driver.execute_script('mobile: tap', x: 0, y: 0, element: element.ref)
+      #   @@driver.execute_async_script('mobile: tap', x: 0, y: 0, element: element.id)
+      #   @@driver.execute_script('mobile: tap', x: 0, y: 0, element: element.id)
       # end
 
       def test_logs