+++ test/unit/script/commands_test.rb
@@ -34,42 +34,11 @@ class ScriptTest
       assert_equal [:get], @c.spec_commands['session/:session_id/element/:id/page_index']
     end
 
-    # depends on webdriver-version... (number of commands)
-    def test_implemented_mjsonwp_commands
-      assert_equal ['sessions', [:get]], @c.implemented_mjsonwp_commands.first
-
-      # pick up an arbitrary command
-      assert_equal %i(get post), @c.implemented_mjsonwp_commands['session/:session_id/alert_text']
-    end
-
-    def test_implemented_w3c_commands
-      # assert_equal 119, @c.implemented_w3c_commands.length
-      assert_equal ['sessions', [:get]], @c.implemented_w3c_commands.first
-
-      # pick up an arbitrary command
-      assert_equal %i(get post), @c.implemented_w3c_commands['session/:session_id/alert/text']
-    end
-
     def test_implemented_core_commands
-      assert_equal ['sessions', [:get]], @c.implemented_core_commands.first
+      assert_equal ['status', [:get]], @c.implemented_core_commands.first
 
       # pick up an arbitrary command
       assert_equal [:post], @c.implemented_core_commands['session/:session_id/appium/device/pull_folder']
-    end
-
-    def test_webdriver_oss_commands
-      assert_equal ['session/:session_id', %i(get delete)], @c.webdriver_oss_commands.first
-
-      # pick up an arbitrary command
-      assert_equal %i(get post), @c.webdriver_oss_commands['session/:session_id/alert_text']
-    end
-
-    def test_webdriver_w3c_commands
-      # assert_equal 47, @c.webdriver_w3c_commands.length
-      assert_equal ['session', [:post]], @c.webdriver_w3c_commands.first
-
-      # pick up an arbitrary command
-      assert_equal %i(get post), @c.webdriver_w3c_commands['session/:session_id/alert/text']
     end
 
     def test_compare_commands_command1_is_bigger
@@ -87,11 +56,7 @@ class ScriptTest
     end
 
     def test_can_call_diff_methods
-      assert_equal({}, @c.all_diff_commands_mjsonwp)
-      assert_equal({}, @c.all_diff_commands_w3c)
       assert_equal({}, @c.diff_except_for_webdriver)
-      assert_equal({}, @c.diff_webdriver_oss)
-      assert_equal({}, @c.diff_webdriver_w3c)
     end
   end
 end