+++ CHANGELOG.md
@@ -10,6 +10,24 @@ Read `release_notes.md` for commit level details.
 
 ### Deprecations
 
+## [5.0.0]
+
+### Enhancements
+
+- Update base selenium webdriver version to `4.0.0`
+  - Support only W3C spec as following Selenium v4 client
+  - Support Ruby 2.6+
+- `element.id` returns the element id instead of `element.ref`. `element.ref` now returns an array.
+- Removed `desired_capabilities` as capabilities for `Appium::Core#for`. Please use `capabilities` key name instead
+- Removals that already had _duplication_ mark
+  - Removed `driver#screenshot`. Please use `driver#save_screenshot` instead
+  - Removed `driver#send_keys` to send keys to an active element. Please use `driver.action.send_keys('happy testing').perform` instead
+  - Removed `forceMjsonwp` to send only MJSONWP capabilities since Selenium cleint v4 no longer supports MJSONWP
+- No longer set default `timeouts` as `0`. ruby_lib_core calls `/timeouts` endpoint only when `appium_lib: { wait: 5 }` is provided explicitly
+- Raises `::Appium::Core::Error::ArgumentError` instead of `ArgumentError` for this library specific argument errors
+- Add `driver#wait`, `driver#wait_until`, `driver#wait_true`, `driver#wait_until_true` syntaxes
+  - Can give `driver` instance as its block variable
+
 ## [4.7.1] - 2021-09-26
 
 ### Enhancements